Add TokenPriceRegistry and DexRegistry setup in deployment scripts
